# Improve FineFundamental backtesting performance - static cache