--- a/Engine/DataFeeds/Enumerators/Factories/FineFundamentalSubscriptionEnumeratorFactory.cs
+++ b/Engine/DataFeeds/Enumerators/Factories/FineFundamentalSubscriptionEnumeratorFactory.cs
@@ -15,6 +15,7 @@
 */
 
 using System;
+using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.Globalization;
 using System.IO;
@@ -33,9 +34,11 @@ namespace QuantConnect.Lean.Engine.DataFeeds.Enumerators.Factories
     /// </summary>
     public class FineFundamentalSubscriptionEnumeratorFactory : ISubscriptionEnumeratorFactory
     {
+        private static readonly ConcurrentDictionary<int, List<DateTime>> FineFilesCache
+            = new ConcurrentDictionary<int, List<DateTime>>();
+
         private readonly bool _isLiveMode;
         private readonly Func<SubscriptionRequest, IEnumerable<DateTime>> _tradableDaysProvider;
-        private DateTime _lastUsedDate;
 
         /// <summary>
         /// Initializes a new instance of the <see cref="FineFundamentalSubscriptionEnumeratorFactory"/> class.
@@ -97,84 +100,77 @@ namespace QuantConnect.Lean.Engine.DataFeeds.Enumerators.Factories
         {
             var source = fine.GetSource(config, date, _isLiveMode);
 
-            if (!File.Exists(source.Source))
+            if (File.Exists(source.Source))
             {
-                if (_lastUsedDate == default(DateTime))
+                return source;
+            }
+
+            var cacheKey = config.Symbol.Value.ToLower().GetHashCode();
+            List<DateTime> availableDates;
+
+            // only use cache in backtest, since in live mode new fine files are added
+            // we still didn't load available fine dates for this symbol
+            if (_isLiveMode || !FineFilesCache.TryGetValue(cacheKey, out availableDates))
+            {
+                try
                 {
-                    // find first file date
-                    List<string> availableFiles;
-                    try
-                    {
-                        availableFiles = Directory.GetFiles(Path.GetDirectoryName(source.Source),
-                            "*.zip").OrderBy(x => x).ToList();
-                    }
-                    catch
-                    {
-                        // argument null exception or directory doesn't exist
-                        return source;
-                    }
-
-                    var fileName = Path.GetFileNameWithoutExtension(source.Source);
-
-                    // no files or requested date before first date, return null source
-                    if (availableFiles.Count == 0
-                        || string.CompareOrdinal(fileName,
-                            Path.GetFileNameWithoutExtension(availableFiles[0])) < 0)
-                    {
-                        return source;
-                    }
-
-                    // take advantage of the order and avoid parsing and comparing all files to get the date
-                    var index = availableFiles.BinarySearch(fileName, new FileNameComparer());
-
-                    if (index < 0)
-                    {
-                        // if negative returns the complement of the index of the next element that is larger than Date
-                        // so subtract 1 to get the previous item which is less than Date
-                        index = ~index - 1;
-                    }
-
-                    var current = DateTime.ParseExact(
-                        Path.GetFileNameWithoutExtension(availableFiles[index]),
-                        "yyyyMMdd",
-                        CultureInfo.InvariantCulture);
-
-                    if (current <= date)
-                    {
-                        // we found the file, save its date and return the source
-                        _lastUsedDate = current;
-                        return fine.GetSource(config, current, _isLiveMode);
-                    }
-
-                    // this shouldn't ever happen so throw
-                    throw new InvalidOperationException("FineFundamentalSubscriptionEnumeratorFactory.GetSource(): " +
-                                                        $"failed to resolve source for {config.Symbol} on {date}");
+                    var path = Path.GetDirectoryName(source.Source) ?? string.Empty;
+                    availableDates = Directory.GetFiles(path, "*.zip")
+                        .Select(
+                            filePath =>
+                            {
+                                try
+                                {
+                                    return DateTime.ParseExact(
+                                        Path.GetFileNameWithoutExtension(filePath),
+                                        "yyyyMMdd",
+                                        CultureInfo.InvariantCulture
+                                    );
+                                }
+                                catch
+                                {
+                                    // just in case...
+                                    return DateTime.MaxValue;
+                                }
+                            }
+                        )
+                        .Where(time => time != DateTime.MaxValue)
+                        .OrderBy(x => x)
+                        .ToList();
                 }
-                else
+                catch
                 {
-                    // return source for last existing file date
-                    source = fine.GetSource(config, _lastUsedDate, _isLiveMode);
+                    // directory doesn't exist or path is null
+                    if (!_isLiveMode)
+                    {
+                        // only add to cache if not live mode
+                        FineFilesCache[cacheKey] = new List<DateTime>();
+                    }
+                    return source;
+                }
+
+                if (!_isLiveMode)
+                {
+                    // only add to cache if not live mode
+                    FineFilesCache[cacheKey] = availableDates;
                 }
             }
-            else
+
+            // requested date before first date, return null source
+            if (availableDates.Count == 0 || date < availableDates[0])
             {
-                _lastUsedDate = date;
+                return source;
+            }
+            for (var i = availableDates.Count - 1; i >= 0; i--)
+            {
+                // we iterate backwards ^ and find the first data point before 'date'
+                if (availableDates[i] <= date)
+                {
+                    return fine.GetSource(config, availableDates[i], _isLiveMode);
+                }
             }
 
             return source;
-        }
-
-        /// <summary>
-        /// Helper class to compare to file names which are named as a date.
-        /// </summary>
-        private class FileNameComparer : IComparer<string>
-        {
-            public int Compare(string x, string y)
-            {
-                return string.CompareOrdinal(
-                    Path.GetFileNameWithoutExtension(x),
-                    Path.GetFileNameWithoutExtension(y));
-            }
         }
     }
 }